13 色々実装

+++ backend/app/domain/model/restaurant.go
@@ -8,5 +8,5 @@ type Restaurant struct {
 	CloseTime time.Time
 	PhotoUrl  string
 	PlaceId   string //dependent to Google Maps API
-	Rating    int
+	Rating    float32
 }
